almost reverted deep inquire from moly kit
