+++ moly-kit/src/widgets.rs
@@ -6,11 +6,9 @@ mod avatar;
 mod chat_lines;
 mod citation;
 mod citation_list;
-mod deep_inquire_line;
 mod message_loading;
 mod message_markdown;
 mod message_thinking_block;
-mod stages;
 
 #[cfg(any(feature = "async-rt", feature = "async-web"))]
 pub mod chat;
@@ -30,8 +28,6 @@ pub fn live_design(cx: &mut makepad_widgets::Cx) {
     message_loading::live_design(cx);
     avatar::live_design(cx);
     chat_lines::live_design(cx);
-    stages::live_design(cx);
-    deep_inquire_line::live_design(cx);
     messages::live_design(cx);
     prompt_input::live_design(cx);
     chat::live_design(cx);
