+++ moly-kit/src/widgets/chat.rs
@@ -165,25 +165,7 @@ impl Chat {
                 MessagesAction::EditSave(index) => {
                     let mut tasks = self.messages_ref().read_with(|m| {
                         let mut message = m.messages[index].clone();
-                        let new_text = m.current_editor_text().expect("no editor text");
-                        
-                        // Update content with new text from editor
-                        match &message.content {
-                            MessageContent::PlainText { citations, .. } => {
-                                message.content = MessageContent::PlainText {
-                                    text: new_text,
-                                    citations: citations.clone(),
-                                };
-                            },
-                            MessageContent::MultiStage { stages, citations, .. } => {
-                                message.content = MessageContent::MultiStage {
-                                    text: new_text,
-                                    stages: stages.clone(),
-                                    citations: citations.clone(),
-                                };
-                            }
-                        }
-                        
+                        message.content.text = m.current_editor_text().expect("no editor text");
                         ChatTask::UpdateMessage(index, message).into()
                     });
 
@@ -194,25 +176,10 @@ impl Chat {
                         let mut messages = m.messages[0..=index].to_vec();
 
                         let index = m.current_editor_index().expect("no editor index");
-                        let new_text = m.current_editor_text().expect("no editor text");
+                        let text = m.current_editor_text().expect("no editor text");
 
-                        // Update content with new text from editor
-                        match &messages[index].content {
-                            MessageContent::PlainText { citations, .. } => {
-                                messages[index].content = MessageContent::PlainText {
-                                    text: new_text,
-                                    citations: citations.clone(),
-                                };
-                            },
-                            MessageContent::MultiStage { stages, citations, .. } => {
-                                messages[index].content = MessageContent::MultiStage {
-                                    text: new_text,
-                                    stages: stages.clone(),
-                                    citations: citations.clone(),
-                                };
-                            }
-                        }
-                        
+                        messages[index].content.text = text;
+
                         vec![ChatTask::SetMessages(messages), ChatTask::Send]
                     });
 
@@ -236,9 +203,9 @@ impl Chat {
                     next_index,
                     Message {
                         from: EntityId::User,
-                        content: MessageContent::PlainText {
-                            text: text.clone(),
-                            citations: Vec::new(),
+                        content: MessageContent {
+                            text,
+                            ..Default::default()
                         },
                         is_writing: false,
                     },
@@ -271,11 +238,8 @@ impl Chat {
 
             messages.messages.push(Message {
                 from: EntityId::Bot(bot_id.clone()),
-                content: MessageContent::PlainText {
-                    text: String::new(),
-                    citations: Vec::new(),
-                },
                 is_writing: true,
+                ..Default::default()
             });
 
             messages
@@ -293,7 +257,9 @@ impl Chat {
         let ui = self.ui_runner();
         let future = async move {
             let mut client = repo.client();
-            let bot = repo.get_bot(&bot_id).expect(format!("No bot found for the given id: {:?}", bot_id).as_str());
+            let bot = repo
+                .get_bot(&bot_id)
+                .expect(format!("No bot found for the given id: {:?}", bot_id).as_str());
 
             let mut message_stream = client.send_stream(&bot, &context);
             while let Some(result) = message_stream.next().await {
@@ -387,8 +353,8 @@ impl Chat {
         match task {
             ChatTask::CopyMessage(index) => {
                 self.messages_ref().read_with(|m| {
-                    let text = m.messages[*index].visible_text();
-                    cx.copy_to_clipboard(&text);
+                    let text = &m.messages[*index].content.text;
+                    cx.copy_to_clipboard(text);
                 });
             }
             ChatTask::DeleteMessage(index) => {
@@ -474,23 +440,17 @@ impl Chat {
         self.prompt_input_ref().write().set_send();
         self.messages_ref().write().messages.retain_mut(|m| {
             m.is_writing = false;
-            // Only remove messages that are completely empty (no content at all)
-            match &m.content {
-                MessageContent::PlainText { text, .. } => !text.is_empty(),
-                MessageContent::MultiStage { stages, .. } => !stages.is_empty() && stages.iter().any(|s| {
-                    s.thinking.is_some() || s.writing.is_some() || s.completed.is_some()
-                }),
-            }
+            !m.content.is_empty()
         });
     }
 
-    fn handle_message_delta(&mut self, cx: &mut Cx, result: ClientResult<MessageDelta>) -> bool {
+    fn handle_message_delta(&mut self, cx: &mut Cx, result: ClientResult<MessageContent>) -> bool {
         let messages = self.messages_ref();
 
         // For simplicity, lets handle this as an standard Result, ignoring delta
         // if there are errors.
         match result.into_result() {
-            Ok(delta) => {
+            Ok(content) => {
                 // Let's abort if we don't have where to put the delta.
                 let Some(mut message) = messages.read().messages.last().cloned() else {
                     return true;
@@ -503,7 +463,7 @@ impl Chat {
                     }
                 }
 
-                message.apply_delta(delta);
+                message.content = content;
 
                 let index = messages.read().messages.len() - 1;
                 let mut tasks = vec![ChatTask::UpdateMessage(index, message)];
@@ -532,9 +492,9 @@ impl Chat {
                             messages.read().messages.len() + i,
                             Message {
                                 from: EntityId::App,
-                                content: MessageContent::PlainText {
+                                content: MessageContent {
                                     text: e.to_string(),
-                                    citations: Vec::new(),
+                                    ..Default::default()
                                 },
                                 is_writing: false,
                             },
