+++ moly-mini/src/tester_client.rs
@@ -18,11 +18,12 @@ impl BotClient for TesterClient {
         &mut self,
         _bot: &Bot,
         messages: &[Message],
-    ) -> MolyStream<'static, ClientResult<MessageDelta>> {
+    ) -> MolyStream<'static, ClientResult<MessageContent>> {
         let mut input = messages
             .last()
             .expect("didn't receive any messages")
-            .visible_text()
+            .content
+            .text
             .split_whitespace()
             .map(|b| b.to_lowercase())
             .collect::<VecDeque<_>>();
@@ -31,11 +32,9 @@ impl BotClient for TesterClient {
             match input.pop_front().as_deref() {
                 Some("say") => {
                     let body = input.make_contiguous().join(" ");
-                    ClientResult::new_ok(MessageDelta {
-                        content: MessageContent::PlainText {
-                            text: body.into(),
-                            citations: vec![],
-                        },
+                    ClientResult::new_ok(MessageContent {
+                        text: body.into(),
+                        ..Default::default()
                     })
                 }
                 Some("error") => ClientResult::new_err(
@@ -54,23 +53,17 @@ impl BotClient for TesterClient {
                     ClientError::new(ClientErrorKind::Unknown, "This is another error".into())
                         .into(),
                 ]),
-                Some("hello") => ClientResult::new_ok(MessageDelta {
-                    content: MessageContent::PlainText {
-                        text: "world".into(),
-                        citations: vec![],
-                    },
+                Some("hello") => ClientResult::new_ok(MessageContent {
+                    text: "world".into(),
+                    ..Default::default()
                 }),
-                Some("ping") => ClientResult::new_ok(MessageDelta {
-                    content: MessageContent::PlainText {
-                        text: "pong".into(),
-                        citations: vec![],
-                    },
+                Some("ping") => ClientResult::new_ok(MessageContent {
+                    text: "pong".into(),
+                    ..Default::default()
                 }),
-                _ => ClientResult::new_ok(MessageDelta {
-                    content: MessageContent::PlainText {
-                        text: "Yeah...".into(),
-                        citations: vec![],
-                    },
+                _ => ClientResult::new_ok(MessageContent {
+                    text: "Yeah...".into(),
+                    ..Default::default()
                 }),
             }
         });
